refactor: merge two files into one
