--- apps/web/src/services/dashboards/components/dashboard-detail/DashboardManageVariableTable.vue
@@ -1,189 +1,0 @@
-<script setup lang="ts">
-import {
-    computed, reactive,
-} from 'vue';
-import type { TranslateResult } from 'vue-i18n';
-
-import {
-    PBadge, PDataTable, PToggleButton, PIconButton,
-} from '@cloudforet/mirinae';
-import type { DataTableField } from '@cloudforet/mirinae/src/data-display/tables/data-table/type';
-
-import type { DashboardGlobalVariable } from '@/schema/dashboard/_types/dashboard-global-variable-type';
-import { i18n } from '@/translations';
-
-import { useDashboardStore } from '@/store/dashboard/dashboard-store';
-
-import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
-
-import ErrorHandler from '@/common/composables/error/errorHandler';
-
-import {
-    MANAGED_DASHBOARD_GLOBAL_VARIABLES_SCHEMA,
-} from '@/services/dashboards/constants/managed-dashboard-global-variables';
-import { useDashboardDetailInfoStore } from '@/services/dashboards/stores/dashboard-detail-info-store';
-
-
-
-interface GlobalVariableTableItem {
-    key: string;
-    name: string;
-    management: VariableType;
-    type?: string;
-    use?: boolean;
-    created_by?: string;
-    reference?: string|TranslateResult;
-}
-interface EmitFn {
-    (e: 'delete', value: string): void;
-    (e: 'edit', name: string): void;
-    (e: 'clone', name: string): void;
-}
-type VariableType = 'managed'|'custom';
-const emit = defineEmits<EmitFn>();
-
-const dashboardStore = useDashboardStore();
-const dashboardDetailStore = useDashboardDetailInfoStore();
-const dashboardDetailState = dashboardDetailStore.state;
-const dashboardDetailGetters = dashboardDetailStore.getters;
-const state = reactive({
-    globalVariablesTableItems: computed<GlobalVariableTableItem[]>(() => {
-        const _managedItems: GlobalVariableTableItem[] = [];
-        Object.values(MANAGED_DASHBOARD_GLOBAL_VARIABLES_SCHEMA).forEach((d) => {
-            _managedItems.push({
-                key: d.key,
-                name: d.name,
-                management: 'managed',
-                use: dashboardDetailGetters.dashboardVarsSchemaProperties[d.key]?.use || false,
-                created_by: 'System',
-                reference: i18n.t('DASHBOARDS.DETAIL.VARIABLES.COMMON'),
-            });
-        });
-        const varsSchemaProperties: DashboardGlobalVariable[] = Object.values(dashboardDetailGetters.dashboardVarsSchemaProperties);
-        const _customItems: GlobalVariableTableItem[] = varsSchemaProperties.map((d) => ({
-            ...d,
-            type: d.method === 'manual' ? d.type : undefined,
-            reference: d.method === 'dynamic' ? d?.reference?.resourceType : undefined,
-        }));
-        return [
-            ..._managedItems,
-            ..._customItems,
-        ];
-    }),
-    variableFields: [
-        { name: 'name', label: 'Name' },
-        { name: 'type', label: 'Type' },
-        { name: 'reference', label: 'Reference' },
-        { name: 'management', label: 'Management' },
-        { name: 'use', label: 'Use', width: '90px' },
-        { name: 'created_by', label: 'Created By' },
-        { name: 'buttons', label: ' ', width: '144px' },
-    ] as DataTableField[],
-    variableType: computed<Partial<Record<VariableType, TranslateResult>>>(() => ({
-        managed: i18n.t('DASHBOARDS.CUSTOMIZE.VARIABLES.FILTER_MANAGED'),
-        custom: i18n.t('DASHBOARDS.CUSTOMIZE.VARIABLES.FILTER_CUSTOM'),
-    })),
-});
-
-
-/* Api */
-const updateDashboardVarsSchema = async (dashboardId: string, variableKey: string, use: boolean) => {
-    try {
-        await dashboardStore.updateDashboard(dashboardId, {
-            dashboard_id: dashboardId,
-            vars_schema: {
-                properties: {
-                    ...dashboardDetailGetters.dashboardVarsSchemaProperties,
-                    [variableKey]: {
-                        ...dashboardDetailGetters.dashboardVarsSchemaProperties[variableKey],
-                        use,
-                    },
-                },
-            },
-        });
-        showSuccessMessage(i18n.t('DASHBOARDS.DETAIL.VARIABLES.ALT_S_UPDATE_DASHBOARD_VARS_SCHEMA'), '');
-    } catch (e) {
-        ErrorHandler.handleRequestError(e, i18n.t('DASHBOARDS.DETAIL.VARIABLES.ALT_E_UPDATE_DASHBOARD_VARS_SCHEMA'));
-    }
-};
-
-/* EVENT */
-const handleCloneVariable = (variableKey: string) => {
-    emit('clone', variableKey);
-};
-const handleEditVariable = (variableKey: string) => {
-    emit('edit', variableKey);
-};
-const handleDeleteVariable = (variableKey: string) => {
-    emit('delete', variableKey);
-};
-const handleToggleUse = (variableKey: string, val: boolean) => {
-    if (!dashboardDetailState.dashboardId) return;
-    updateDashboardVarsSchema(dashboardDetailState.dashboardId, variableKey, val);
-};
-
-/* Helper */
-const variableTypeBadgeStyleFormatter = (type: VariableType) => {
-    if (type === 'managed') return 'gray500';
-    return 'primary';
-};
-</script>
-
-<template>
-    <div class="list-wrapper">
-        <p-data-table class="variable-table"
-                      :items="state.globalVariablesTableItems"
-                      :fields="state.variableFields"
-                      :loading="dashboardDetailState.loadingDashboard"
-        >
-            <template #col-type-format="{ value }">
-                {{ value || '-' }}
-            </template>
-            <template #col-reference-format="{ value }">
-                {{ value || '-' }}
-            </template>
-            <template #col-management-format="{ value }">
-                <p-badge :style-type="variableTypeBadgeStyleFormatter(value)"
-                         badge-type="solid-outline"
-                >
-                    {{ state.variableType[value] }}
-                </p-badge>
-            </template>
-            <template #col-use-format="{ value, item }">
-                <p-toggle-button :value="value"
-                                 :disabled="item.disabled || item.required || item.fixed || item.readonly"
-                                 @change-toggle="handleToggleUse(item.key, $event)"
-                />
-            </template>
-            <template #col-buttons-format="{ item }">
-                <div class="button-wrapper">
-                    <p-icon-button name="ic_clone"
-                                   size="sm"
-                                   @click="handleCloneVariable(item.key)"
-                    />
-                    <p-icon-button v-if="item.management === 'custom'"
-                                   name="ic_edit"
-                                   size="sm"
-                                   @click="handleEditVariable(item.key)"
-                    />
-                    <p-icon-button v-if="item.management === 'custom'"
-                                   name="ic_delete"
-                                   size="sm"
-                                   style-type="negative-transparent"
-                                   @click="handleDeleteVariable(item.key)"
-                    />
-                </div>
-            </template>
-        </p-data-table>
-    </div>
-</template>
-
-<style lang="postcss" scoped>
-.list-wrapper {
-    .variable-table {
-        .button-wrapper {
-            @apply w-full flex items-center;
-        }
-    }
-}
-</style>
